Added the 'bytes per second' info to NetworkInfo
I added the calculation in WindowsMonitorService, because the information i needed to do it was easily accessable there.
Can be moved to the class that populates NetworkInfo but that would take a bit more work.

Hopefully the code can survive adding af another netcard like a wifi dongle or removing one while it is running.